Moved XMLFetchTask from MainActivity to SplashActivity
