+++ VideoCollectionTemplate/source/VideoCollectionTemplateApp/app/src/main/java/org/buildmlearn/videocollection/activities/MainActivity.java
@@ -1,15 +1,12 @@
 package org.buildmlearn.videocollection.activities;
 
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.widget.Toast;
 
-import org.buildmlearn.videocollection.Constants;
 import org.buildmlearn.videocollection.NetworkUtils;
 import org.buildmlearn.videocollection.R;
-import org.buildmlearn.videocollection.data.FetchXMLTask;
 import org.buildmlearn.videocollection.fragment.MainActivityFragment;
 
 /**
@@ -17,29 +14,15 @@ import org.buildmlearn.videocollection.fragment.MainActivityFragment;
  */
 public class MainActivity extends AppCompatActivity implements MainActivityFragment.Callback {
 
-    private SharedPreferences prefs = null;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        prefs = getSharedPreferences(Constants.firstrun, MODE_PRIVATE);
-
         if (NetworkUtils.isNetworkAvailable(this)) {
             Toast.makeText(this, R.string.network_unavailable, Toast.LENGTH_SHORT).show();
         }
 
-    }
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-
-        if (prefs.getBoolean(Constants.firstrun, true)) {
-            FetchXMLTask xmlTask = new FetchXMLTask(this);
-            xmlTask.execute(Constants.XMLFileName);
-            prefs.edit().putBoolean(Constants.firstrun, false).apply();
-        }
     }
 
     @Override
